6.15 OpCodes (corrected)
Ran through OpcodeWizard and rechecked all the opcodes now that it's operational again.

Also put opcodes in the same order as opcode wizard for easier reading experiences.

I've added the game version this matches against so that anyone who wants to compare asset version to game version can now check if they match. This field will be updated to ffxivgame.ver as needed.